requirements.txt
arrow==1.2.3
binaryornot==0.4.4
build==0.9.0
cached-property==1.5.2
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
cookiecutter==2.1.1
coverage==6.5.0
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
packaging==21.3
pep517==0.13.0
pyfakefs==5.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.13
